Power energy values from config

+++ src/api/java/com/enderio/api/capacitor/QuadraticScalable.java
@@ -8,7 +8,7 @@ import java.util.function.Supplier;
  * A value that is scaled using a quadratic equation.
  * (base, level) => base * level^2
  */
-public record QuadraticScalable(CapacitorModifier modifier, Supplier<Float> base) implements ICapacitorScalable {
+public record QuadraticScalable(CapacitorModifier modifier, Supplier<Integer> base) implements ICapacitorScalable {
 
     @Override
     public Supplier<Float> scaleF(NonNullSupplier<ICapacitorData> data) {
@@ -20,7 +20,7 @@ public record QuadraticScalable(CapacitorModifier modifier, Supplier<Float> base
         return () -> Math.round(scale(base.get(), data.get().getModifier(modifier)));
     }
 
-    private static float scale(float base, float level) {
+    private static float scale(int base, float level) {
         return base * level * level;
     }
 }
